This is part of ongoing work for multi platform support

This PR brings Linux x86_x64 compilation

* windows structures are now controlled by us
* most structures have the basic foundations for future 32bit
abstractions
* explicit 16bit unicode support (Linux is 32bit Unicode)
* update of all syscall code to use the new structures and unicode
system

runtime support of the emulator under linux is still ongoing.

+++ src/windows-emulator/memory_utils.hpp
@@ -21,7 +21,7 @@ inline std::string get_permission_string(const memory_permission permission)
 
 inline memory_permission map_nt_to_emulator_protection(uint32_t nt_protection)
 {
-	nt_protection &= ~PAGE_GUARD; // TODO: Implement that
+	nt_protection &= ~static_cast<uint32_t>(PAGE_GUARD); // TODO: Implement that
 
 	switch (nt_protection)
 	{
